requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.4.1
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.9
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
coverage==6.2
cryptography==38.0.3
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.2.0
Django==4.0
django-cleanup==6.0.0
django-cors-headers==3.10.1
django-environ==0.9.0
django-import-export==3.0.2
django-ip-geolocation==1.6.1
django-ipware==4.0.2
django-jazzmin==2.6.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
drf-extra-fields==3.4.1
drf-yasg==1.21.4
et-xmlfile==1.1.0
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==15.3.4
frozenlist==1.3.3
future==0.18.2
geoip2==4.6.0
html5lib==1.1
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
install==1.3.5
itypes==1.2.0
Jinja2==3.0.3
lxml==4.9.1
MarkupPy==1.14
MarkupSafe==2.0.1
maxminddb==2.2.0
multidict==6.0.4
numpy==1.21.5
odfpy==1.4.1
opencv-python==4.5.5.62
openpyxl==3.0.10
oscrypto==1.3.0
packaging==22.0
Pillow==9.3.0
pluggy==1.0.0
psycopg2-binary==2.9.2
pycparser==2.21
pyHanko==0.15.1
pyhanko-certvalidator==0.19.6
PyJWT==2.3.0
PyPDF3==1.0.6
pytest==7.2.0
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.26.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.2
svglib==1.4.1
tablib==3.3.0
tinycss2==1.2.1
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2021.5
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.7
webencodings==0.5.1
whitenoise==6.3.0
xhtml2pdf==0.2.8
xlrd==2.0.1
xlwt==1.3.0
yarl==1.8.2
